Assistants/main  - FE code cleanup (#306)
FE cleanup
