+++ src/interfaces/coral_web/src/components/ConversationError.tsx
@@ -4,6 +4,7 @@ import { isNotFoundError } from '@/cohere-client';
 import { Header } from '@/components/Conversation/Header';
 import { Icon } from '@/components/Shared/Icon';
 import { Text } from '@/components/Shared/Text';
+import { useSlugRoutes } from '@/hooks/slugRoutes';
 import { cn } from '@/utils';
 
 type Props = {
@@ -11,6 +12,10 @@ type Props = {
 };
 
 export const ConversationError: React.FC<Props> = ({ error }) => {
+  const { agentId } = useSlugRoutes();
+
+  const url = agentId ? `/a/${agentId}` : '/';
+
   return (
     <>
       <Header />
@@ -21,7 +26,7 @@ export const ConversationError: React.FC<Props> = ({ error }) => {
           {isNotFoundError(error) ? (
             <>
               This conversation does not exist, <br /> why not create a{' '}
-              <Link href="/" className="underline">
+              <Link href={url} className="underline">
                 new one
               </Link>
               ?
