+++ src/interfaces/coral_web/src/components/Settings/ToolsTab.tsx
@@ -2,12 +2,14 @@ import React, { useMemo } from 'react';
 
 import { ManagedTool } from '@/cohere-client';
 import { ToolsInfoBox } from '@/components/Settings/ToolsInfoBox';
-import { Text } from '@/components/Shared';
+import { Button, Icon, Text } from '@/components/Shared';
 import { ToggleCard } from '@/components/ToggleCard';
 import { WelcomeGuideTooltip } from '@/components/WelcomeGuideTooltip';
 import { TOOL_FALLBACK_ICON, TOOL_ID_TO_DISPLAY_INFO } from '@/constants';
+import { useAgent } from '@/hooks/agents';
 import { useDefaultFileLoaderTool } from '@/hooks/files';
-import { useListTools } from '@/hooks/tools';
+import { useSlugRoutes } from '@/hooks/slugRoutes';
+import { useListTools, useUnauthedTools } from '@/hooks/tools';
 import { useFilesStore, useParamsStore } from '@/stores';
 import { ConfigurableParams } from '@/stores/slices/paramsSlice';
 import { cn } from '@/utils';
@@ -15,43 +17,33 @@ import { cn } from '@/utils';
 /**
  * @description Tools tab content that shows a list of available tools and files
  */
-export const ToolsTab: React.FC<{ className?: string }> = ({ className = '' }) => {
+export const ToolsTab: React.FC<{ requiredTools: string[] | undefined; className?: string }> = ({
+  requiredTools,
+  className = '',
+}) => {
+  const { agentId } = useSlugRoutes();
+  const { data: agent } = useAgent({ agentId });
   const { params, setParams } = useParamsStore();
   const { data } = useListTools();
   const { tools: paramTools } = params;
+  const enabledTools = paramTools ?? [];
   const { defaultFileLoaderTool } = useDefaultFileLoaderTool();
   const { clearComposerFiles } = useFilesStore();
 
-  const { availableTools, unavailableTools } = useMemo(() => {
-    return (data ?? [])
-      .filter((t) => t.is_visible)
-      .reduce<{ availableTools: ManagedTool[]; unavailableTools: ManagedTool[] }>(
-        (acc, tool) => {
-          if (tool.is_available) {
-            acc.availableTools.push(tool);
-          } else {
-            acc.unavailableTools.push(tool);
-          }
-          return acc;
-        },
-        { availableTools: [], unavailableTools: [] }
-      );
-  }, [data]);
-  const enabledTools = paramTools
-    ? availableTools.filter((t) => paramTools.map((t) => t.name).includes(t.name))
-    : [];
+  const { unauthedTools } = useUnauthedTools();
+  const availableTools = useMemo(() => {
+    return (data ?? []).filter(
+      (t) =>
+        t.is_visible &&
+        t.is_available &&
+        (!requiredTools || requiredTools.some((rt) => rt === t.name))
+    );
+  }, [data, requiredTools]);
 
   const handleToggle = (name: string, checked: boolean) => {
-    let tool = availableTools.find((tool) => tool.name === name);
-    if (tool === undefined) {
-      return;
-    }
-    if (tool.is_auth_required && tool.auth_url !== null) {
-      window.location.assign(tool.auth_url!);
-    }
     const newParams: Partial<ConfigurableParams> = {
       tools: checked
-        ? [...enabledTools, tool]
+        ? [...enabledTools, { name }]
         : enabledTools.filter((enabledTool) => enabledTool.name !== name),
     };
 
@@ -68,35 +60,24 @@ export const ToolsTab: React.FC<{ className?: string }> = ({ className = '' }) =
       <ToolsInfoBox />
       <article className={cn('flex flex-col gap-y-5 pb-10')}>
         <Text styleAs="p-sm" className="text-secondary-800">
-          Tools are data sources the assistant can search such as databases or the internet.
+          {availableTools.length === 0
+            ? `${agent?.name} does not use any tools.`
+            : 'Tools are data sources the assistant can search such as databases or the internet.'}
         </Text>
 
-        {unavailableTools.length > 0 && (
+        {unauthedTools.length > 0 && (
           <>
-            <Text as="span" styleAs="label" className="font-medium">
-              Action Required
-            </Text>
-
-            <div className="flex flex-col gap-y-5">
-              {unavailableTools.map(({ name, display_name, description, error_message }) => {
-                return (
-                  <ToggleCard
-                    key={name}
-                    disabled
-                    errorMessage={error_message}
-                    checked={false}
-                    label={display_name ?? name}
-                    icon={TOOL_ID_TO_DISPLAY_INFO[name]?.icon ?? TOOL_FALLBACK_ICON}
-                    description={description ?? ''}
-                    onToggle={(checked) => handleToggle(name, checked)}
-                  />
-                );
-              })}
+            <div className="flex items-center justify-between">
+              <Text as="span" styleAs="label" className="font-medium">
+                Action Required
+              </Text>
+              <Icon name="warning" kind="outline" />
             </div>
+            <ConnectDataBox tools={unauthedTools} />
           </>
         )}
 
-        {unavailableTools.length > 0 && availableTools.length > 0 && (
+        {unauthedTools.length > 0 && availableTools.length > 0 && (
           <hr className="border-t border-marble-400" />
         )}
 
@@ -110,17 +91,22 @@ export const ToolsTab: React.FC<{ className?: string }> = ({ className = '' }) =
               {availableTools.map(({ name, display_name, description, error_message }) => {
                 const enabledTool = enabledTools.find((enabledTool) => enabledTool.name === name);
                 const checked = !!enabledTool;
+                const disabled = !!requiredTools;
 
                 return (
-                  <ToggleCard
-                    key={name}
-                    errorMessage={error_message}
-                    checked={checked}
-                    label={display_name ?? name}
-                    icon={TOOL_ID_TO_DISPLAY_INFO[name]?.icon ?? TOOL_FALLBACK_ICON}
-                    description={description ?? ''}
-                    onToggle={(checked) => handleToggle(name, checked)}
-                  />
+                  <>
+                    <ToggleCard
+                      key={name}
+                      disabled={disabled}
+                      errorMessage={error_message}
+                      checked={checked}
+                      label={display_name ?? name}
+                      name={name}
+                      icon={TOOL_ID_TO_DISPLAY_INFO[name]?.icon ?? TOOL_FALLBACK_ICON}
+                      description={description ?? ''}
+                      onToggle={(checked) => handleToggle(name, checked)}
+                    />
+                  </>
                 );
               })}
             </div>
@@ -129,5 +115,35 @@ export const ToolsTab: React.FC<{ className?: string }> = ({ className = '' }) =
       </article>
       <WelcomeGuideTooltip step={2} className="fixed right-0 mr-3 mt-12 md:right-full md:mt-0" />
     </section>
+  );
+};
+
+/**
+ * @description Info box that prompts the user to connect their data to enable tools
+ */
+const ConnectDataBox: React.FC<{
+  tools: ManagedTool[];
+}> = ({ tools }) => {
+  return (
+    <div className="flex flex-col gap-y-4 rounded border border-dashed border-primary-400 bg-primary-200 p-4">
+      <div className="flex flex-col gap-y-3">
+        <Text styleAs="h5">Connect your data</Text>
+        <Text>
+          In order to get the most accurate answers grounded on your data, connect the following:
+        </Text>
+      </div>
+      <div className="flex flex-col gap-y-1">
+        {tools.map((tool) => (
+          <Button
+            key={tool.name}
+            kind="secondary"
+            href={tool.auth_url ?? ''}
+            endIcon={<Icon name="arrow-up-right" className="ml-1" />}
+          >
+            {tool.display_name}
+          </Button>
+        ))}
+      </div>
+    </div>
   );
 };
