+++ src/interfaces/coral_web/src/hooks/actions.ts
@@ -89,7 +89,7 @@ export const useChatHotKeys = (): CustomHotKey[] => {
       name: 'Start a new conversation',
       commands: ['ctrl+shift+o', 'meta+shift+o'],
       action: async () => {
-        await router.push('/', undefined, { shallow: true });
+        router.push('/', undefined, { shallow: true });
         resetConversation();
         resetCitations();
         resetFileParams();
