--- src/interfaces/coral_web/src/pages/c/[id]/index.tsx
@@ -1,228 +1,0 @@
-import { Transition } from '@headlessui/react';
-import { DehydratedState, QueryClient, dehydrate } from '@tanstack/react-query';
-import { GetServerSideProps, NextPage } from 'next';
-import { useRouter } from 'next/router';
-import { useContext, useEffect } from 'react';
-
-import { CohereClient, Document } from '@/cohere-client';
-import { AgentsList } from '@/components/Agents/AgentsList';
-import { Layout, LeftSection, MainSection } from '@/components/Agents/Layout';
-import Conversation from '@/components/Conversation';
-import { ConversationError } from '@/components/ConversationError';
-import ConversationListPanel from '@/components/ConversationList/ConversationListPanel';
-import { Spinner } from '@/components/Shared';
-import { TOOL_PYTHON_INTERPRETER_ID } from '@/constants';
-import { BannerContext } from '@/context/BannerContext';
-import { useIsDesktop } from '@/hooks/breakpoint';
-import { useConversation } from '@/hooks/conversation';
-import { useListAllDeployments } from '@/hooks/deployments';
-import { useExperimentalFeatures } from '@/hooks/experimentalFeatures';
-import { appSSR } from '@/pages/_app';
-import {
-  useCitationsStore,
-  useConversationStore,
-  useParamsStore,
-  useSettingsStore,
-} from '@/stores';
-import { OutputFiles } from '@/stores/slices/citationsSlice';
-import { cn, getQueryString } from '@/utils';
-import { createStartEndKey, mapHistoryToMessages } from '@/utils';
-import { parsePythonInterpreterToolFields } from '@/utils/tools';
-
-type Props = {
-  reactQueryState: DehydratedState;
-};
-
-const ConversationPage: NextPage<Props> = () => {
-  const router = useRouter();
-  const {
-    params: { deployment },
-    setParams,
-    resetFileParams,
-  } = useParamsStore();
-  const { setConversation } = useConversationStore();
-  const { addCitation, resetCitations, saveOutputFiles } = useCitationsStore();
-  const { data: experimentalFeatures } = useExperimentalFeatures();
-  const isLangchainModeOn = !!experimentalFeatures?.USE_EXPERIMENTAL_LANGCHAIN;
-  const { setMessage } = useContext(BannerContext);
-  const {
-    settings: { isConvListPanelOpen, isMobileConvListPanelOpen },
-  } = useSettingsStore();
-  const isDesktop = useIsDesktop();
-  const isMobile = !isDesktop;
-
-  const urlConversationId = getQueryString(router.query.id);
-
-  const {
-    data: conversation,
-    isLoading,
-    isError,
-    error,
-  } = useConversation({
-    conversationId: urlConversationId,
-  });
-  const { data: allDeployments } = useListAllDeployments();
-
-  useEffect(() => {
-    if (!deployment && allDeployments) {
-      const firstAvailableDeployment = allDeployments.find((d) => d.is_available);
-      if (firstAvailableDeployment) {
-        setParams({ deployment: firstAvailableDeployment.name });
-      }
-    }
-  }, [deployment]);
-
-  useEffect(() => {
-    if (!isLangchainModeOn) return;
-    setMessage('You are using an experimental langchain multihop flow. There will be bugs.');
-  }, [isLangchainModeOn]);
-
-  useEffect(() => {
-    resetCitations();
-    resetFileParams();
-    setParams({ tools: [] });
-
-    if (urlConversationId) {
-      setConversation({ id: urlConversationId });
-    }
-  }, [urlConversationId, setConversation, resetCitations]);
-
-  useEffect(() => {
-    if (!conversation) return;
-
-    const messages = mapHistoryToMessages(
-      conversation?.messages?.sort((a, b) => a.position - b.position)
-    );
-
-    setConversation({ name: conversation.title, messages });
-
-    let documentsMap: { [documentId: string]: Document } = {};
-    let outputFilesMap: OutputFiles = {};
-
-    (conversation?.messages ?? []).forEach((message) => {
-      message.documents?.forEach((doc) => {
-        const docId = doc.document_id ?? '';
-        documentsMap[docId] = doc;
-
-        const toolName = (doc.tool_name ?? '').toLowerCase();
-
-        if (toolName === TOOL_PYTHON_INTERPRETER_ID) {
-          const { outputFile } = parsePythonInterpreterToolFields(doc);
-
-          if (outputFile) {
-            outputFilesMap[outputFile.filename] = {
-              name: outputFile.filename,
-              data: outputFile.b64_data,
-              documentId: docId,
-            };
-          }
-        }
-      });
-      message.citations?.forEach((citation) => {
-        const startEndKey = createStartEndKey(citation.start ?? 0, citation.end ?? 0);
-        const documents = citation.document_ids?.map((id) => documentsMap[id]) ?? [];
-        addCitation(message.generation_id ?? '', startEndKey, documents);
-      });
-    });
-
-    saveOutputFiles(outputFilesMap);
-  }, [conversation?.id, setConversation]);
-
-  return (
-    <Layout>
-      <LeftSection>
-        <AgentsList />
-      </LeftSection>
-
-      <MainSection>
-        <div className="flex h-full">
-          <Transition
-            as="section"
-            show={(isMobileConvListPanelOpen && isMobile) || (isConvListPanelOpen && isDesktop)}
-            enterFrom="translate-x-full lg:translate-x-0 lg:w-0"
-            enterTo="translate-x-0 lg:w-[300px]"
-            leaveFrom="translate-x-0 lg:w-[300px]"
-            leaveTo="translate-x-full lg:translate-x-0 lg:w-0"
-            className={cn(
-              'z-main-section flex lg:min-w-0',
-              'absolute h-full w-full lg:static lg:h-auto',
-              'border-0 border-marble-400 md:border-r',
-              'transition-[transform, width] duration-500 ease-in-out'
-            )}
-          >
-            <ConversationListPanel />
-          </Transition>
-          <Transition
-            as="main"
-            show={isDesktop || !isMobileConvListPanelOpen}
-            enterFrom="-translate-x-full"
-            enterTo="translate-x-0"
-            leaveFrom="translate-x-0"
-            leaveTo="-translate-x-full"
-            className={cn(
-              'flex min-w-0 flex-grow flex-col',
-              'transition-transform duration-500 ease-in-out'
-            )}
-          >
-            {isLoading ? (
-              <div className="flex h-full flex-grow flex-col items-center justify-center">
-                <Spinner />
-              </div>
-            ) : isError ? (
-              <ConversationError error={error} />
-            ) : (
-              <Conversation conversationId={urlConversationId} />
-            )}
-          </Transition>
-        </div>
-      </MainSection>
-    </Layout>
-  );
-};
-
-export const getServerSideProps: GetServerSideProps = async (context) => {
-  const deps = appSSR.initialize() as {
-    queryClient: QueryClient;
-    cohereClient: CohereClient;
-  };
-
-  const conversationId = context.params?.id as string;
-
-  await Promise.allSettled([
-    deps.queryClient.prefetchQuery({
-      queryKey: ['conversation', conversationId],
-      queryFn: async () => {
-        const conversation = await deps.cohereClient.getConversation({
-          conversationId,
-        });
-        // react-query useInfiniteQuery expected response shape
-        return { conversation };
-      },
-    }),
-    deps.queryClient.prefetchQuery({
-      queryKey: ['conversations'],
-      queryFn: async () => {
-        const conversations = await deps.cohereClient.listConversations({});
-        return conversations;
-      },
-    }),
-    deps.queryClient.prefetchQuery({
-      queryKey: ['tools'],
-      queryFn: async () => await deps.cohereClient.listTools({}),
-    }),
-    deps.queryClient.prefetchQuery({
-      queryKey: ['deployments'],
-      queryFn: async () => await deps.cohereClient.listDeployments({}),
-    }),
-  ]);
-
-  return {
-    props: {
-      appProps: {
-        reactQueryState: dehydrate(deps.queryClient),
-      },
-    },
-  };
-};
-
-export default ConversationPage;
