+++ src/interfaces/coral_web/src/pages/new.tsx
@@ -4,7 +4,7 @@ import { GetServerSideProps, NextPage } from 'next';
 import { CohereClient } from '@/cohere-client';
 import { AgentsList } from '@/components/Agents/AgentsList';
 import { CreateAgentForm } from '@/components/Agents/CreateAgentForm';
-import { Layout, LeftSection, MainSection } from '@/components/Agents/Layout';
+import { Layout, LeftSection, MainSection } from '@/components/Layout';
 import { appSSR } from '@/pages/_app';
 
 type Props = {};
